Support slot sizes larger than 2GB and add a max_slot_size option (#126)
* Keep the active message for reliable subscribers

A reliable subscriber now retains an internal reference to its current
active message even when keep_active_message is false.  Without it the
slot could be recycled by the publisher while the message was still the
subscriber's current message.

LockWeakMessage() uses the same predicate so that locking a weak_ptr for
the current message returns the retained message instead of counting the
slot a second time against max_active_messages.

Adds docs/reliable-messages.md describing slot lifetime, publisher
backpressure, the effect of a slow subscriber and active message
retention, plus a test with two reliable subscribers that checks every
message is delivered to both.

* Update version

* Make slot sizes 64 bit

Slot size was an int32_t in PublisherOptions, in the CreatePublisher
wire protocol and in ServerChannel, so the largest usable slot was
2GB - 64 bytes.  Anything larger wrapped negative and the server
rejected it with "num_slots and slot_size must be greater than 0".
INT32_MAX itself also failed, because the client 64-byte aligns the
request before sending it and the aligned value overflowed the field.

Widen slot size to int64_t through the options, the protobuf messages,
the client and server channels, the growth path and the C, Python, Rust
and Java bindings.  GetMessageBuffer() takes an int64_t max size, so the
auto-growth loop no longer truncates as it approaches 2GB.

ChannelControlBlock::max_message_size becomes a uint64_t to match; it
was silently truncating messages over 4GB.  The Rust client mirrors this
struct, so both now assert the layout of the stats block.

The Android ByteBuffer capacity is still an int, so the JNI layer
reports an error rather than returning an undefined buffer for a slot
that a JVM cannot represent.

* Select the client API slot size width at build time

Slot sizes stay 64 bit internally: the wire protocol, the shared memory
layout and all of the buffer arithmetic are unchanged.  The C and C++
client APIs, however, now go back to exposing a 32 bit slot size by
default so that code written against the older API keeps compiling.

subspace::SlotSizeType and the C SubspaceSlotSize typedef select the
width.  Define SUBSPACE_64BIT_SLOT_SIZE to widen them and allow slots
larger than 2GB to be requested and reported, via --config=slot_size_64
for Bazel or -DSUBSPACE_64BIT_SLOT_SIZE=ON for CMake.  Like the existing
shared memory backend option, it changes struct layouts and function
signatures, so it has to be set consistently for the library and for
everything that includes its headers.

The Python, Rust and Java clients are unconditionally 64 bit.

The test that needs a slot larger than 2GB is compiled out of the
default build, so CI gains a slot-size-64 job to cover the widened API.

* Document the 64 bit slot size build option

* Add a max_slot_size publisher option to cap channel growth

* Report whether the large slot test ran in CI

* Update version to 3.2.0

+++ android/java/com/subspace/SubspaceClient.java
@@ -50,7 +50,7 @@ public class SubspaceClient implements AutoCloseable {
      * @return a new publisher
      * @throws SubspaceException on failure
      */
-    public SubspacePublisher createPublisher(String channelName, int slotSize,
+    public SubspacePublisher createPublisher(String channelName, long slotSize,
                                              int numSlots, boolean reliable) {
         checkOpen();
         long pubHandle = nativeCreatePublisher(nativeHandle, channelName,
@@ -59,7 +59,7 @@ public class SubspaceClient implements AutoCloseable {
     }
 
     /** Create a non-reliable publisher with default options. */
-    public SubspacePublisher createPublisher(String channelName, int slotSize,
+    public SubspacePublisher createPublisher(String channelName, long slotSize,
                                              int numSlots) {
         return createPublisher(channelName, slotSize, numSlots, false);
     }
@@ -97,7 +97,7 @@ public class SubspaceClient implements AutoCloseable {
     private static native void nativeDestroy(long handle);
     private static native long nativeCreatePublisher(long handle,
                                                      String channelName,
-                                                     int slotSize,
+                                                     long slotSize,
                                                      int numSlots,
                                                      boolean reliable);
     private static native long nativeCreateSubscriber(long handle,
