+++ rpc/common/rpc_common.h
@@ -13,21 +13,21 @@
 namespace subspace {
 
 // Slot parameters for requests and responses for the open/close requests.
-constexpr int32_t kRpcRequestSlotSize = 128;
-constexpr int32_t kRpcResponseSlotSize = 128;
+constexpr int64_t kRpcRequestSlotSize = 128;
+constexpr int64_t kRpcResponseSlotSize = 128;
 constexpr int32_t kRpcRequestNumSlots = 100;
 constexpr int32_t kRpcResponseNumSlots = 100;
 
 // Default slot parameters for method invocation requests.  The slot size
 // can be expanded if the request is bigger.
-constexpr int32_t kDefaultMethodSlotSize = 256;
+constexpr int64_t kDefaultMethodSlotSize = 256;
 constexpr int32_t kDefaultMethodNumSlots = 100;
 
-constexpr int32_t kCancelChannelSlotSize = 64;
+constexpr int64_t kCancelChannelSlotSize = 64;
 constexpr int32_t kCancelChannelNumSlots = 8;
 
 struct MethodOptions {
-  int32_t slot_size = kDefaultMethodSlotSize;
+  int64_t slot_size = kDefaultMethodSlotSize;
   int32_t num_slots = kDefaultMethodNumSlots;
   int id = -1;
 };
@@ -39,7 +39,7 @@ struct Method {
   int id;
   std::string request_type;
   std::string response_type;
-  int32_t slot_size;
+  int64_t slot_size;
   int32_t num_slots;
   std::shared_ptr<subspace::Publisher> request_publisher;
   std::shared_ptr<subspace::Subscriber> response_subscriber;
